semop.2, proc.5, cgroup_namespaces.7, cgroups.7: Fix scrambled scopes
The manpages commonly use .in +4/.EX....EE/.in pairs for iindented examples.

Clean up four exceptional instances out of hundreds in which the .in
was the inner operation rather than the outer. No difference in
rendering, but the uniformity makes it possible to parse these
mechanically with one rule.

Really all these ought to be replaced with .RS/.EX....EE/.RE but
doclifter can cope given this less intrusive change.