proc.5: Setting dumpable to 1 reverts ownership of /proc/PID/* to effective IDs
The statement that resetting the dumpable attribute of a process
to 1 causes the ownership of files to revert the process's real
IDs looked suspect. And indeed it is at odds with the code in
fs/proc/base.c::task_dump_owner() (Linux 4.16 sources).
Further verified with a quick test that resetting dumpable to 1
causes the ownership of /proc/PID/* files to revert to the
process's effective IDs. Mea culpa for the original mistake.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>